bot/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
aiosqlite==0.19.0
aiofiles==23.1.0
pydantic==1.10.13
email-validator>=2.0.0
aiohttp>=3.8.0,<4.0.0
python-dotenv>=1.0.0
slowapi==0.1.9
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart>=0.0.18
sentry-sdk[fastapi]>=2.15.0

